backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Realtime Note Editor with AI"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "alembic>=1.16.5",
    "fastapi[all]>=0.116.1",
    "langchain>=0.3.27",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "sentry-sdk[fastapi]>=2.35.2",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
